requirements.txt
appdirs==1.4.4
argh==0.26.2
asgiref==3.4.1
atomicwrites==1.4.0
boto==2.49.0
brotlipy==0.7.0
certifi==2021.10.8
click==8.0.3
clyent==1.2.2
cycler==0.10.0
cytoolz==0.11.0
daal4py==2021.3.0
dask==2021.10.0
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
et-xmlfile==1.1.0
fonttools==4.25.0
gmpy2==2.0.8
llvmlite==0.37.0
locket==0.2.1
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
navigator-updater==0.2.1
nltk==3.6.5
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==8.4.0
pkginfo==1.7.1
ply==3.11
pycosat==0.6.3
pycurl==7.44.1
pyls-spyder==0.4.0
pyodbc===4.0.0-unsupported
pytest==6.2.4
python-dotenv==0.19.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
pytz==2021.3
PyYAML==6.0
simplegeneric==0.8.1
sip==4.19.13
Sphinx==4.2.0
sqlparse==0.4.2
TBB==0.2
terminado==0.9.4
urllib3==1.26.7
webencodings==0.5.1
xlwt==1.3.0
zict==2.0.0
zope.event==4.5.0

